+++ index.html
@@ -165,7 +165,7 @@
     </section>
   </template>
 
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
 
 </body>
 </html>
