+++ js/util.js
@@ -1,0 +1,18 @@
+const getRandomPositiveInteger = (a, b) => {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+
+const getRandomElementFromArray = (items) => items[Math.floor(Math.random() * items.length)];
+
+
+const checkStringLength = (string, length) => string.length <= length;
+
+
+export { getRandomPositiveInteger, getRandomElementFromArray, checkStringLength };
